other update

+++ module2/practic2_2/calculator.c
@@ -1,7 +1,7 @@
-int sum(int a, int b) { return a + b; }
+double sum(double a, double b) { return a + b; }
 
-int sub(int a, int b) { return a - b; }
+double sub(double a, double b) { return a - b; }
 
-int divide(int a, int b) { return a / b; }
+double divide(double a, double b) { return a / b; }
 
-int mul(int a, int b) { return a * b; }
+double mul(double a, double b) { return a * b; }
